renamed host_resolver_t.h and host_resolver_t.cpp to not have "_t"
