+++ src/installer/corehost/cli/apphost/static/CMakeLists.txt
@@ -22,13 +22,13 @@ include_directories(../../json)
 
 set(SOURCES
     ../bundle_marker.cpp
-    ./hostfxr_resolver_t.cpp
+    ./hostfxr_resolver.cpp
     ./hostpolicy_resolver.cpp
 )
 
 set(HEADERS
     ../bundle_marker.h
-    ../../../hostfxr_resolver_t.h
+    ../../../hostfxr_resolver.h
 )
 
 if(CLR_CMAKE_TARGET_WIN32)
