+++ src/installer/corehost/hostfxr_resolver.h
